requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
pydantic==2.11.7
python-multipart==0.0.20
jinja2==3.1.6
openpyxl==3.1.5
itsdangerous==2.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.16.4
pytest==8.4.1
httpx==0.28.1
python-dotenv==1.0.1
